+++ requirements.txt
@@ -5,3 +5,4 @@ scipy>=0.9
 sklearn
 passlib>=1.6
 bcrypt
+pyyaml
